+++ CSClock/About.Designer.cs
@@ -60,6 +60,7 @@ namespace CSClock
             this.textBoxDescription = new System.Windows.Forms.TextBox();
             this.okButton = new System.Windows.Forms.Button();
             this.button_licenses = new System.Windows.Forms.Button();
+            this.button1 = new System.Windows.Forms.Button();
             this.tableLayoutPanel.SuspendLayout();
             ((System.ComponentModel.ISupportInitialize)(this.logoPictureBox)).BeginInit();
             this.SuspendLayout();
@@ -228,11 +229,22 @@ namespace CSClock
             this.button_licenses.UseVisualStyleBackColor = true;
             this.button_licenses.Click += new System.EventHandler(this.button1_Click);
             // 
+            // button1
+            // 
+            this.button1.Location = new System.Drawing.Point(24, 397);
+            this.button1.Name = "button1";
+            this.button1.Size = new System.Drawing.Size(387, 23);
+            this.button1.TabIndex = 7;
+            this.button1.Text = "Submit feedback / Donate";
+            this.button1.UseVisualStyleBackColor = true;
+            this.button1.Click += new System.EventHandler(this.button1_Click_1);
+            // 
             // About
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.ClientSize = new System.Drawing.Size(435, 400);
+            this.ClientSize = new System.Drawing.Size(435, 429);
+            this.Controls.Add(this.button1);
             this.Controls.Add(this.button_licenses);
             this.Controls.Add(this.tableLayoutPanel);
             this.Controls.Add(this.button_openGitHubPage);
@@ -272,5 +284,6 @@ namespace CSClock
         private System.Windows.Forms.TextBox textBoxDescription;
         private System.Windows.Forms.Button okButton;
         private System.Windows.Forms.Button button_licenses;
+        private System.Windows.Forms.Button button1;
     }
 }