+++ CSClock/DebugMenu.Designer.cs
@@ -48,6 +48,7 @@ namespace CSClock
         {
             this.label1 = new System.Windows.Forms.Label();
             this.button1 = new System.Windows.Forms.Button();
+            this.button2 = new System.Windows.Forms.Button();
             this.SuspendLayout();
             // 
             // label1
@@ -70,15 +71,28 @@ namespace CSClock
             this.button1.UseVisualStyleBackColor = true;
             this.button1.Click += new System.EventHandler(this.button1_Click);
             // 
+            // button2
+            // 
+            this.button2.Font = new System.Drawing.Font("Microsoft Sans Serif", 14.25F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+            this.button2.Location = new System.Drawing.Point(12, 274);
+            this.button2.Name = "button2";
+            this.button2.Size = new System.Drawing.Size(401, 47);
+            this.button2.TabIndex = 2;
+            this.button2.Text = "Close the debug menu";
+            this.button2.UseVisualStyleBackColor = true;
+            this.button2.Click += new System.EventHandler(this.button2_Click);
+            // 
             // DebugMenu
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.ClientSize = new System.Drawing.Size(425, 263);
+            this.ClientSize = new System.Drawing.Size(425, 333);
+            this.Controls.Add(this.button2);
             this.Controls.Add(this.button1);
             this.Controls.Add(this.label1);
             this.Name = "DebugMenu";
             this.Text = "Debug menu - CSClock";
+            this.TopMost = true;
             this.ResumeLayout(false);
 
         }
@@ -87,5 +101,6 @@ namespace CSClock
 
         private System.Windows.Forms.Label label1;
         private System.Windows.Forms.Button button1;
+        private System.Windows.Forms.Button button2;
     }
 }