+++ CSClock/DebugMenu.cs
@@ -39,5 +39,10 @@ namespace CSClock
         {
             Program.CheckForUpdate(true);
         }
+
+        private void button2_Click(object sender, EventArgs e)
+        {
+            Close();
+        }
     }
 }