+++ CSClock/Properties/Settings.Designer.cs
@@ -12,7 +12,7 @@ namespace CSClock.Properties {
     
     
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.Editors.SettingsDesigner.SettingsSingleFileGenerator", "14.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.Editors.SettingsDesigner.SettingsSingleFileGenerator", "15.1.0.0")]
     internal sealed partial class Settings : global::System.Configuration.ApplicationSettingsBase {
         
         private static Settings defaultInstance = ((Settings)(global::System.Configuration.ApplicationSettingsBase.Synchronized(new Settings())));
@@ -329,6 +329,29 @@ namespace CSClock.Properties {
             }
             set {
                 this["portableAutoUpdaterQuestionShown"] = value;
+            }
+        }
+        
+        [global::System.Configuration.UserScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        public global::System.DateTime startDate {
+            get {
+                return ((global::System.DateTime)(this["startDate"]));
+            }
+            set {
+                this["startDate"] = value;
+            }
+        }
+        
+        [global::System.Configuration.UserScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        [global::System.Configuration.DefaultSettingValueAttribute("False")]
+        public bool feedbackReminderShown {
+            get {
+                return ((bool)(this["feedbackReminderShown"]));
+            }
+            set {
+                this["feedbackReminderShown"] = value;
             }
         }
     }